+++ lib/librato/rails.rb
@@ -5,6 +5,7 @@ require_relative 'rails/version'
 
 # must load after all module setup and in this order
 if defined?(Rails)
+  require_relative 'rails/version_specifier'
   require_relative 'rails/railtie'
   require_relative 'rails/subscribers'
   require_relative 'rails/helpers'