+++ test/dummy/extras/dummy_job.rb
@@ -1,0 +1,7 @@
+class DummyJob < ActiveJob::Base
+  queue_as :default
+
+  def perform
+    sleep 1
+  end
+end